Transform addons reducer/actions/action creators into a ducks module
