+++ tests/unit/amo/components/TestAddonReviewList.js
@@ -12,8 +12,7 @@ import {
 import Link from 'amo/components/Link';
 import Paginate from 'core/components/Paginate';
 import { loadEntities } from 'core/actions';
-import { fetchAddon } from 'core/actions/addons';
-import { denormalizeAddon } from 'core/reducers/addons';
+import { denormalizeAddon, fetchAddon } from 'core/reducers/addons';
 import ErrorList from 'ui/components/ErrorList';
 import Rating from 'ui/components/Rating';
 import { fakeAddon, fakeReview } from 'tests/unit/amo/helpers';
