+++ tests/unit/core/sagas/testAddons.js
@@ -1,10 +1,9 @@
 import { hideLoading, showLoading } from 'react-redux-loading-bar';
 import SagaTester from 'redux-saga-tester';
 
-import { fetchAddon } from 'core/actions/addons';
 import * as api from 'core/api';
 import { ENTITIES_LOADED } from 'core/constants';
-import addonsReducer from 'core/reducers/addons';
+import addonsReducer, { fetchAddon } from 'core/reducers/addons';
 import apiReducer from 'core/reducers/api';
 import addonsSaga from 'core/sagas/addons';
 import { dispatchSignInActions, fakeAddon } from 'tests/unit/amo/helpers';
